+++ prometheus/http_test.go
@@ -14,9 +14,6 @@
 package prometheus
 
 import (
-	"bytes"
-	"errors"
-	"log"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -121,114 +118,4 @@ func TestInstrumentHandler(t *testing.T) {
 	if want, got := 1., out.Counter.GetValue(); want != got {
 		t.Errorf("want reqCnt of %f, got %f", want, got)
 	}
-}
-
-type errorCollector struct{}
-
-func (e errorCollector) Describe(ch chan<- *Desc) {
-	ch <- NewDesc("invalid_metric", "not helpful", nil, nil)
-}
-
-func (e errorCollector) Collect(ch chan<- Metric) {
-	ch <- NewInvalidMetric(
-		NewDesc("invalid_metric", "not helpful", nil, nil),
-		errors.New("collect error"),
-	)
-}
-
-func TestHandlerErrorHandling(t *testing.T) {
-
-	// Create a registry that collects a MetricFamily with two elements,
-	// another with one, and reports an error.
-	reg := NewRegistry()
-
-	cnt := NewCounter(CounterOpts{
-		Name: "the_count",
-		Help: "Ah-ah-ah! Thunder and lightning!",
-	})
-	reg.MustRegister(cnt)
-
-	cntVec := NewCounterVec(
-		CounterOpts{
-			Name:        "name",
-			Help:        "docstring",
-			ConstLabels: Labels{"constname": "constvalue"},
-		},
-		[]string{"labelname"},
-	)
-	cntVec.WithLabelValues("val1").Inc()
-	cntVec.WithLabelValues("val2").Inc()
-	reg.MustRegister(cntVec)
-
-	reg.MustRegister(errorCollector{})
-
-	logBuf := &bytes.Buffer{}
-	logger := log.New(logBuf, "", 0)
-
-	writer := httptest.NewRecorder()
-	request, _ := http.NewRequest("GET", "/", nil)
-	request.Header.Add("Accept", "test/plain")
-
-	errorHandler := HandlerFor(reg, HandlerOpts{
-		ErrorLog:      logger,
-		ErrorHandling: HTTPErrorOnError,
-	})
-	continueHandler := HandlerFor(reg, HandlerOpts{
-		ErrorLog:      logger,
-		ErrorHandling: ContinueOnError,
-	})
-	panicHandler := HandlerFor(reg, HandlerOpts{
-		ErrorLog:      logger,
-		ErrorHandling: PanicOnError,
-	})
-	wantMsg := `error collecting metrics: 1 error(s) occurred:
-
-* error collecting metric Desc{fqName: "invalid_metric", help: "not helpful", constLabels: {}, variableLabels: []}: collect error
-`
-	wantErrorBody := `An error has occurred during metrics collection:
-
-1 error(s) occurred:
-
-* error collecting metric Desc{fqName: "invalid_metric", help: "not helpful", constLabels: {}, variableLabels: []}: collect error
-`
-	wantOKBody := `# HELP name docstring
-# TYPE name counter
-name{constname="constvalue",labelname="val1"} 1
-name{constname="constvalue",labelname="val2"} 1
-# HELP the_count Ah-ah-ah! Thunder and lightning!
-# TYPE the_count counter
-the_count 0
-`
-
-	errorHandler.ServeHTTP(writer, request)
-	if got, want := writer.Code, http.StatusInternalServerError; got != want {
-		t.Errorf("got HTTP status code %d, want %d", got, want)
-	}
-	if got := logBuf.String(); got != wantMsg {
-		t.Errorf("got log message %q, want %q", got, wantMsg)
-	}
-	if got := writer.Body.String(); got != wantErrorBody {
-		t.Errorf("got body %q, want %q", got, wantErrorBody)
-	}
-	logBuf.Reset()
-	writer.Body.Reset()
-	writer.Code = http.StatusOK
-
-	continueHandler.ServeHTTP(writer, request)
-	if got, want := writer.Code, http.StatusOK; got != want {
-		t.Errorf("got HTTP status code %d, want %d", got, want)
-	}
-	if got := logBuf.String(); got != wantMsg {
-		t.Errorf("got log message %q, want %q", got, wantMsg)
-	}
-	if got := writer.Body.String(); got != wantOKBody {
-		t.Errorf("got body %q, want %q", got, wantOKBody)
-	}
-
-	defer func() {
-		if err := recover(); err == nil {
-			t.Error("expected panic from panicHandler")
-		}
-	}()
-	panicHandler.ServeHTTP(writer, request)
 }
